+++ source/config/structure.js
@@ -55,10 +55,20 @@ let type =
         if (!(value instanceof RegExp || typeof value === 'string'))
             return `${name} must be a regular expression or string`;
     },
+    regexParams: (value, name) =>
+    {
+        if (!(Array.isArray(value) || value instanceof RegExp || typeof value === 'string'))
+            return `${name} must be an array, a regular expression, or a string`;
+    },
     string: (value, name) =>
     {
         if (typeof value !== 'string')
             return `${name} must be a string`;
+    },
+    stringOrFunction: (value, name) =>
+    {
+        if (!(typeof value === 'string' || typeof value === 'function'))
+            return `${name} must be a string or a function`;
     },
     stringArray: (value, name) =>
     {
@@ -127,6 +137,16 @@ export default
                 [$validate]: type.stringArray,
 
                 '{*}': type.stringArray
+            }
+        },
+
+        delousers:
+        {
+            '{*}': 
+            {
+                [$validate]: type.stringArray,
+                search: type.regexParams,
+                replace: type.stringOrFunction
             }
         },